Server: Refactor server to improve readability
Change-type: patch
