+++ lib/index.js
@@ -1,3 +1,3 @@
-const createServer = require('./server.js')
+const createServer = require('./server')
 
 createServer()
